// src/reducers/user/types.ts
enum USER_ACTION_TYPES {
  SET_CURRENT_USER = 'user/SET_CURRENT_USER',
  CHECK_USER_SESSION = 'user/CHECK_USER_SESSION',
  GOOGLE_SIGN_IN_REQUEST = 'user/GOOGLE_SIGN_IN_REQUEST',
  EMAIL_SIGN_IN_REQUEST = 'user/EMAIL_SIGN_IN_REQUEST',
  SIGN_IN_SUCCESS = 'user/SIGN_IN_SUCCESS',
  SIGN_IN_FAILURE = 'user/SIGN_IN_FAILURE',
  SIGN_UP_REQUEST = 'user/SIGN_UP_REQUEST',
  SIGN_UP_SUCCESS = 'user/SIGN_UP_SUCCESS',
  SIGN_UP_FAILURE = 'user/SIGN_UP_FAILURE',
  SIGN_OUT_REQUEST = 'user/SIGN_OUT_REQUEST',
  SIGN_OUT_SUCCESS = 'user/SIGN_OUT_SUCCESS',
  SIGN_OUT_FAILURE = 'user/SIGN_OUT_FAILURE',
}

export default USER_ACTION_TYPES;